chore(app): Extract clickable help text widget
